Load map data using maps::map()

load_maps_data(): Wrapper function for maps::map() that loads map data for a specified country or countries, optionally transforms it to WGS 84 format, and returns an sf object.

Usage

load_maps_data(
  countries = c("Sweden", "Denmark", "Russia", "Estonia", "Latvia", "Lithuania",
    "Finland"),
  wgs84 = TRUE
)

Arguments

countries

A character vector of countries for which to load data, default is c("Sweden", "Denmark", "Russia", "Estonia", "Latvia", "Lithuania", "Finland") intended for a map of Norway

wgs84

Logical, whether to transform the data to WGS 84 (EPSG:4326), default is TRUE

Value

An sf object

See also

Examples

norway <- load_maps_data(countries = "Norway")
europe <- load_maps_data()